move types into its own folders
