+++ backend/src/entities/users/validator.ts
@@ -3,11 +3,7 @@ import { createInsertSchema, createUpdateSchema } from "drizzle-zod";
 import z from "zod";
 import { Expo } from "expo-server-sdk";
 import { MIN_LIMIT } from "../../constants/database";
-import { PaginationParams, paginationSchema } from "../../utilities/pagination";
-
-export type CreateUserPayload = typeof usersTable.$inferInsert;
-export type UpdateUserPayload = Partial<typeof usersTable.$inferInsert>;
-export type User = typeof usersTable.$inferSelect;
+import { paginationSchema } from "../../utilities/pagination";
 
 export const createUserValidate = createInsertSchema(usersTable, {
   name: (schema) => schema.min(MIN_LIMIT),
@@ -24,26 +20,6 @@ export const expoTokenValidate = z.object({
     message: "Invalid Expo Push Token",
   }),
 });
-
-export interface SearchedUser {
-  id: string;
-  name: string;
-  username: string;
-  isMember: boolean;
-  profilePhoto: string | null;
-}
-
-export interface SearchedInfo {
-  groupId: string;
-  username: string;
-  limit: number;
-  page: number;
-  userId: string;
-}
-
-export type Pagination = PaginationParams & {
-  id: string;
-};
 
 export const querySchema = z
   .object({
