+++ backend/src/types/api/internal/members.ts
@@ -1,8 +1,9 @@
-import { membersTable } from "../schema";
+import { membersTable } from "../../../entities/schema";
 
 export type AddMemberPayload = {
   userId: string;
   groupId: string;
   role: "MEMBER" | "MANAGER";
 };
+
 export type Member = typeof membersTable.$inferSelect;
